backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.6
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.3
